pyproject.toml
[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"


[project]
name = "FediVuln"
version = "1.1.0"
description = "A client to gather vulnerability-related information from the Fediverse."
authors = [
    {name = "Cédric Bonhomme", email = "cedric.bonhomme@circl.lu"}
]
license = "GPL-3.0-or-later"
readme = "README.md"
keywords = ["Vulnerability-Lookup", "Vulnerability", "CVE", "Fediverse", "Mastodon"]

dynamic = ["classifiers"]

requires-python = ">=3.10,<4.0"
dependencies = [
    "pyvulnerabilitylookup (>=2.2.0)",
    "mastodon-py (>=1.8.1)",
    "valkey (>=6.0.2)"
]


[project.urls]
Homepage = "https://github.com/vulnerability-lookup/FediVuln"
Repository = "https://github.com/vulnerability-lookup/FediVuln"
Changelog = "https://github.com/vulnerability-lookup/FediVuln/blob/main/CHANGELOG.md"


[project.scripts]
FediVuln-Register = "fedivuln.register:main"
FediVuln-Stream = "fedivuln.stream:main"
FediVuln-Search = "fedivuln.search:main"
FediVuln-Publish = "fedivuln.publish:main"


[tool.poetry]
requires-poetry = ">=2.0"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Security",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"
]
include = [
    "README.md",
    "COPYING",
    "fedivuln/conf_sample.py",
]


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
flake8 = "^6.0.0"
pre-commit = "^4.0.1"
types-requests = "^2.32.0.20241016"


[tool.poetry.group.dev]
optional = true


[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
strict_optional = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true

show_error_context = true
pretty = true

exclude = "build|dist|docs"

[tool.isort]
profile = "black"
